Changed Repository unit tests into integration tests. (+1 squashed commits)
Squashed commits:

[f40ac9f] Expanded repository functionality, added tests

+++ GameRater.DAL/Models/Game.cs
@@ -11,7 +11,7 @@ namespace GameRater.DAL.Models
         public int Id { get; set; }
         public string Title { get; set; }
         public string Summary { get; set; }
-        public int ReviewScore { get; set; }
+        public double ReviewScore { get; set; }
 
         public List<Review> Reviews { get; set; }
     }
